Poprawki przed szkoleniem
